Enabled the robot and its workspace to be drawn in the same plot
Modified the draw_tdcr function and added ax parameter to enable working with external figures. Functions the same as before if ax is not provided. \nModified the workspace method to show the plot only if ax parameter is not provided. If so, uses the external plot to populate the workspace.